+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The `FixedWidthUnsignedInteger` trait is now sealed and cannot be implemented outside of `agb-fixnum`.
 - The `num!` macro can now be used in const contexts and is forced to be evaluated at compile time.
+- `TileSet` no longer has a generic parameter, and is always assumed to be `'static`.
 
 ## [0.22.6] - 2025/10/29